requirements.txt
# ====================================================
# ODIN-AI Python Dependencies
# Created: 2025-09-24
# Python 3.8+ Required
# ====================================================

# ===== Core Framework =====
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-dotenv==1.0.0
python-multipart==0.0.6

# ===== Database =====
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1
asyncpg==0.29.0

# ===== Redis Cache =====
redis==5.0.1
aioredis==2.0.1

# ===== API & HTTP =====
httpx==0.25.2
aiohttp==3.9.1
requests==2.31.0
beautifulsoup4==4.12.2

# ===== Document Processing =====
# HWP 처리 (Python 3.13과 호환되는 버전)
# pyhwp는 Python 2.7용이므로 제외, 대신 hwp5 사용
olefile==0.47

# PDF 처리
PyPDF2==3.0.1
pdfplumber==0.10.3

# Excel 처리
openpyxl==3.1.2
pandas==2.1.4
xlrd==2.0.1

# Markdown 변환
markdown==3.5.1
html2text==2020.1.16

# ===== AI & NLP =====
openai==1.3.7
tiktoken==0.5.2
langchain==0.0.348
chromadb==0.4.19

# ===== Task Queue =====
celery==5.3.4
redis==5.0.1
flower==2.0.1

# ===== Email =====
python-email-validator==2.1.0
jinja2==3.1.2

# ===== Security =====
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
bcrypt==4.1.2

# ===== Logging & Monitoring =====
loguru==0.7.2
sentry-sdk==1.39.1

# ===== Testing =====
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
faker==20.1.0

# ===== Development Tools =====
black==23.12.0
flake8==6.1.0
mypy==1.7.1
pre-commit==3.6.0
ipython==8.18.1

# ===== Date & Time =====
python-dateutil==2.8.2
pytz==2023.3.post1

# ===== Utilities =====
pyyaml==6.0.1
click==8.1.7
tqdm==4.66.1
python-slugify==8.0.1
validators==0.22.0

# ===== Async Support =====
# asyncio는 Python 표준 라이브러리에 포함되어 있음
aiofiles==23.2.1

# ===== Data Validation =====
marshmallow==3.20.1
marshmallow-sqlalchemy==0.29.0

# ===== Image Processing (Optional) =====
Pillow==10.1.0

# ===== Compression =====
python-magic==0.4.27